rename to Coins.h to avoid conflicts
